requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular
psycopg2
psycopg2-binary
django-filter
django-simple-history